requirements.txt
beautifultable==1.1.0
coloredlogs==15.0.1
comet_ml==3.31.21
functorch==1.13.1
hydra-core==1.3.2
matplotlib==3.6.2
numpy==1.23.4
omegaconf==2.2.3
pandas==1.5.1
Pillow==9.4.0
pytest==7.2.1
torch==1.13.1
torchmetrics==0.10.3
torchvision==0.14.1
tqdm==4.64.1
wandb==0.13.11
tensorboard==2.12.0
gpustat==1.0.0
mkdocs==1.4.2
mkdocs-material==9.1.2
mkdocstrings==0.20.0
mkdocstrings-python==0.8.3
jax
chex==0.1.5
flax==0.6.7
optax==0.1.4
jax_resnet==0.0.4